Added initial neural net code.

+++ cqi_cpp/include/qtreetwo.hpp
@@ -4,9 +4,9 @@
 class QTreeTwo : public QTree {
     public:
         QTree* qtree;
-        unordered_map<int, float>* action_map;
+        unordered_map<int, double>* action_map;
         vector<LeafSubpolicy*>* subpolicies;
 
-        QTreeTwo(QTree*, unordered_map<int, float>*); // Subpolicies are generated in constructor
-        void trainSubpolicies();                
+        QTreeTwo(QTree*, unordered_map<int, double>*); // Subpolicies are generated in constructor
+        void trainSubpolicies(); // Calls train() on all LeafSubpolicies                
 };
